major: Revamp repo and pip package structure
The structure of the respository was quite legacy. Here it is updated to a
more modern pip package that should be easier to maintain.

Most of the changes are cosmetic, the functionality didn't change. Still, the
API is now slightly different as the GoogleDriveDownloader class has been
replaced by a simple function. Since the logic is stateless, there was no
reason to have this class around.

The only package dependency, i.e. requests, has been added to the toml config
so that it will be finally handled via pip at installation time (see:
https://github.com/ndrplz/google-drive-downloader/issues/29).

The examples in the README have been updated as the previous links were not
working anymore.
