+++ README.md
@@ -2,23 +2,25 @@
 Minimal class to download shared files from Google Drive.
 
 ### How to install:
-Installing is as simple as
+
+This package is on [PyPI](https://pypi.org/project/googledrivedownloader/). Installing is as simple as
 
 ```
 pip install googledrivedownloader
 ```
 
 ### Hello World
-You will need to obtain the sharable link from Google Drive:
 
 ```python
-from google_drive_downloader import GoogleDriveDownloader as gdd
+from googledrivedownloader import download_file_from_google_drive
 
-gdd.download_file_from_google_drive(file_id='1iytA1n2z4go3uVCwE__vIKouTKyIDjEq',
-                                    dest_path='./data/mnist.zip',
-                                    unzip=True)
+# Single image file
+download_file_from_google_drive(file_id='1H1ett7yg-TdtTt6mj2jwmeGZaC8iY1CH',
+                                dest_path='data/crossing.jpg')
+# Zip archive
+download_file_from_google_drive(file_id='13nD8T7_Q9fkQzq9bXF2oasuIZWao8uio',
+                                dest_path='data/docs.zip', unzip=True)
 ```
-This will download a `mnist.zip` file into a `data` folder and unzip it.
 
 
 ### Tips
