--- google_drive_downloader/__init__.py
@@ -1,1 +1,0 @@
-from .google_drive_downloader import GoogleDriveDownloader
