--- setup.cfg
@@ -1,5 +1,0 @@
-[metadata]
-description-file = README.md
-
-[bdist_wheel]
-universal=1
